source/requirements.txt
aliyun-python-sdk-core-v3==2.9.4
aliyun-python-sdk-iot==4.1.1
aliyunsdkcore==1.0.3
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
kombu==4.2.1
Pillow==5.3.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.6.6
PyMySQL==0.9.2
pytz==2018.5
rabbitmq==0.2.0
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.12
tornado==5.1.1
urllib3==1.23
vine==1.1.4
pypinyin==0.33.2
openpyxl==2.5.9
